improve type inference in JSON N5Reader.listAttributes
integer numbers are interpreted as long
arrays of integer numbers are interpreted as long[]
arrays of numbers that include at least one non-integer numbers are
interpreted as double[]

Note that this is weak inference because JSON does not distinguish
integer and floating point numbers.  It also leads to double
attributes that have integer values to be interpreted as long.

TODO consider String based inference to enable "1.0" doubles

arrays of mixed types are interpreted as Object[]

I consider this a bugfix