feat: add and use KeyValuAccess path compose method
abstracts further to platform dependent separation strings
(e.g. "/" vs "\"). The resulting path string is now strictly a platform
dependent path and should not be used in code where a platform
independent string is expected.  Thanks to the Windows FileSystem
implementation accepting "/" separated paths, this should not cause
pracitcal issues at this time but we may have to change the API to support
other FileSystems (e.g. passing paths along as components instead of
strings or settling on a nio Path like object).