fix: assumption that the first deepListHelper result is always self
It seems that on some systems (e.g. github actions) the order of the deepListHelper results
is not gauranteed such that the first entry is always self.

When it is not, we throw away a needed result, and later get an error as we try to
process a result which *is* self, and is an empty string. This causes the
substring method to throw an error, though generally it is swallowed later,
and rarely seen (another problem)