fix: DataBlock decoding
The old, wrong behavior limits the length of the input stream
containing the compressed data to be the length of the uncompressed data.

Rather, the whole input stream should be decompressed. Any splitting
or limiting of the input ReadData should happen before it is passed to
a DataBlockCodec.

This caused problems for blosc specifically