Add N5Writer.writeRegion methods
writeRegion() takes a DataBlockSupplier and a region in pixel
coordinates that should be written. It decised which shards/blocks need
to be touched and asks the DataBlockSupplier for the respective blocks.

Currently only minimal parallelization (over shards) is implemented.
This could be improved in the future, if called with a ForkJoinPool
executor. Then we could further parallelize the production and encoding
of DataBlocks.